stock-analysis/requirements.txt
# Perplexity SDK (preferred)
perplexityai>=0.1.0

# OpenAI SDK (fallback compatibility mode)
openai>=1.0.0

# Agent Framework (LangChain 1.0+)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# Note: OpenRouter is used via OpenAI-compatible API (no additional package needed)

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging
structlog>=23.0.0

